Refactor: 불필요한 코드 제거 및 혈당카드뷰 페이지 분리
